Consolidate MCP server module
