+++ pyproject.toml
@@ -66,7 +66,7 @@ dev = [
 gpu = ["faiss-gpu>=1.7.0"]
 
 [project.scripts]
-pycontextify = "pycontextify.mcp_server:main"
+pycontextify = "pycontextify.mcp:main"
 
 [project.urls]
 Homepage = "https://github.com/pycontextify/pycontextify"
